refactor(e2e): move single-node tests into multi-node/single-node on SingleNodeTestContext
Relocate the 14 single-node-topology tests into multi-node/single-node/ and switch
them onto SingleNodeTestContext. Apply the FAST_REORG_TIMING preset to l1_reorgs and
the six optimistic_proving reorg blocks; drop now-redundant explicit inboxLag: 2.
Merge the proving trio (multiple + empty_blocks_proof + long_proving_time) into
single-node/proving.parallel.test.ts (per-it setup, keeping multiple's world-state-
prune assertion as its own it). Co-locate partial_proof (its unique startProof path
preserved) with partial_proof_multi_root, and manual_rollback with sync_after_reorg
(node_reorg_recovery.test.ts, flattening the redundant nested describe). Update
.test_patterns.yml + bootstrap.sh CI discovery for the new subfolder (proving.parallel
keeps the 15m timeout the long-proving scenario needs).
