--- yarn-project/end-to-end/src/multi-node/empty_blocks_proof.test.ts
@@ -1,55 +1,0 @@
-import type { Logger } from '@aztec/aztec.js/log';
-import { RollupContract } from '@aztec/ethereum/contracts';
-import { ChainMonitor } from '@aztec/ethereum/test';
-import { sleep } from '@aztec/foundation/sleep';
-
-import { jest } from '@jest/globals';
-
-import type { EndToEndContext } from '../fixtures/utils.js';
-import { MultiNodeTestContext } from './multi_node_test_context.js';
-
-jest.setTimeout(1000 * 60 * 15);
-
-// Single-node epoch suite (default MultiNodeTestContext, no extra validator nodes). Starts a prover
-// node (fake proofs). Sets minTxsPerBlock=1 after setup so blocks are empty, then verifies that
-// the prover still submits a proof for those empty-block checkpoints within the proof submission
-// window.
-describe('multi-node/empty_blocks_proof', () => {
-  let context: EndToEndContext;
-  let rollup: RollupContract;
-  let logger: Logger;
-  let monitor: ChainMonitor;
-
-  let L1_BLOCK_TIME_IN_S: number;
-
-  let test: MultiNodeTestContext;
-
-  beforeEach(async () => {
-    test = await MultiNodeTestContext.setup({});
-    ({ context, rollup, logger, monitor, L1_BLOCK_TIME_IN_S } = test);
-  });
-
-  afterEach(async () => {
-    jest.restoreAllMocks();
-    await test.teardown();
-  });
-
-  // Raises minTxsPerBlock to 1 so the sequencer cannot build blocks, advances to epoch 1,
-  // then waits for the prover to submit a proof for the empty checkpoint. Asserts that the
-  // monitor's checkpointNumber matches the proven target, confirming the proof landed on L1.
-  it('submits proof even if there are no txs to build a block', async () => {
-    context.sequencer?.updateConfig({ minTxsPerBlock: 1 });
-    await test.waitUntilEpochStarts(1);
-
-    // REFACTOR: raw sleep to flush pending L1 txs; replace with a helper that waits for the
-    // sequencer to finish all in-flight L1 publishes (e.g. waitForSequencerIdle).
-    // Sleep to make sure any pending checkpoints are published
-    await sleep(L1_BLOCK_TIME_IN_S * 1000);
-    const checkpointNumberAtEndOfEpoch0 = await rollup.getCheckpointNumber();
-    logger.info(`Starting epoch 1 after checkpoint ${checkpointNumberAtEndOfEpoch0}`);
-
-    await test.waitUntilProvenCheckpointNumber(checkpointNumberAtEndOfEpoch0, 240);
-    expect(monitor.checkpointNumber).toEqual(checkpointNumberAtEndOfEpoch0);
-    logger.info(`Test succeeded`);
-  });
-});
