+++ yarn-project/end-to-end/src/multi-node/single-node/partial_proof_multi_root.test.ts
@@ -6,8 +6,9 @@ import { EpochTestSettler } from '@aztec/aztec/testing';
 import { MAX_CHECKPOINTS_PER_EPOCH } from '@aztec/constants';
 import { OutboxContract, type ViemL2ToL1Msg } from '@aztec/ethereum/contracts';
 import type { L1ContractAddresses } from '@aztec/ethereum/l1-contract-addresses';
+import type { ChainMonitor } from '@aztec/ethereum/test';
 import type { ExtendedViemWalletClient } from '@aztec/ethereum/types';
-import { EpochNumber } from '@aztec/foundation/branded-types';
+import { CheckpointNumber, EpochNumber } from '@aztec/foundation/branded-types';
 import { retryUntil } from '@aztec/foundation/retry';
 import { OutboxAbi } from '@aztec/l1-artifacts';
 import { TestContract } from '@aztec/noir-test-contracts.js/Test';
@@ -23,7 +24,7 @@ import { type TxReceipt, TxStatus } from '@aztec/stdlib/tx';
 import { jest } from '@jest/globals';
 import { type Hex, decodeEventLog } from 'viem';
 
-import { MultiNodeTestContext } from './multi_node_test_context.js';
+import { SingleNodeTestContext } from '../single_node_test_context.js';
 
 jest.setTimeout(1000 * 60 * 10);
 
@@ -31,13 +32,13 @@ jest.setTimeout(1000 * 60 * 10);
 // manually to stage progressively deeper partial-proof roots (K=1, 2, 3) for the same epoch, then
 // asserts: (a) the node picks the smallest covering root, (b) any covering root produces a valid
 // consume tx, (c) the shared bitmap blocks double-spend, and (d) K=4 can be staged later.
-// MultiNodeTestContext: single node, no prover, prod-seq, interval mining. Timing: ethSlot=default
+// SingleNodeTestContext: single node, no prover, prod-seq, interval mining. Timing: ethSlot=default
 // (8s/12s CI), aztecSlot=default, epoch=1000, proofSubmissionEpochs=1024 (v5: the disableAnvilTestWatcher
 // override was removed and a perBlockAllocationMultiplier=1.3 was added so the first block of the
 // now-up-to-5-block checkpoint has enough DA budget for the TestContract deploy tx). The test actively
 // calls the Outbox L1 contract to consume L2-to-L1 messages → cross-chain.
-describe('multi-node/partial_proof_multi_root', () => {
-  let test: MultiNodeTestContext;
+describe('multi-node/single-node/partial_proof_multi_root', () => {
+  let test: SingleNodeTestContext;
   let logger: Logger;
   let node: AztecNode;
   let l1Client: ExtendedViemWalletClient;
@@ -50,7 +51,7 @@ describe('multi-node/partial_proof_multi_root', () => {
   let recipient: EthAddress;
 
   beforeEach(async () => {
-    test = await MultiNodeTestContext.setup({
+    test = await SingleNodeTestContext.setup({
       numberOfAccounts: 1,
       minTxsPerBlock: 1,
       // With the enforced timetable this setup can have 5 blocks per checkpoint. The default
@@ -373,4 +374,39 @@ describe('multi-node/partial_proof_multi_root', () => {
     expect(decoded.args.messageHash).toBe(leaf.toString());
     expect(decoded.args.leafId).toBe(getL2ToL1MessageLeafId(witness));
   }
+});
+
+// Co-located with the multi-root suite above: both manually drive partial-epoch proving on a
+// single node with a very long epoch. This one is the only coverage of the prover-node `startProof`
+// path (the multi-root suite hand-drives an EpochTestSettler with no prover), so it stays its own
+// `describe` with its own setup rather than folding into the multi-root `beforeEach`.
+describe('multi-node/single-node/partial_proof', () => {
+  let logger: Logger;
+  let monitor: ChainMonitor;
+
+  let test: SingleNodeTestContext;
+
+  beforeEach(async () => {
+    test = await SingleNodeTestContext.setup({ aztecEpochDuration: 1000 });
+    ({ monitor, logger } = test);
+  });
+
+  afterEach(async () => {
+    jest.restoreAllMocks();
+    await test.teardown();
+  });
+
+  // Waits for 4 checkpoints to land, then calls proverNode.startProof(epoch=0) and polls
+  // ChainMonitor.provenCheckpointNumber until it exceeds 0, confirming a partial proof was
+  // accepted on-chain.
+  it('submits partial proofs when instructed manually', async () => {
+    // With pipelining, each checkpoint takes ~2 L2 slots on a solo-sequencer setup.
+    await test.waitUntilCheckpointNumber(CheckpointNumber(4), test.L2_SLOT_DURATION_IN_S * 12);
+    logger.info(`Kicking off partial proof`);
+
+    await test.context.proverNode!.getProverNode()!.startProof(EpochNumber(0));
+    await test.waitUntilProvenCheckpointNumber(CheckpointNumber(1));
+
+    logger.info(`Test succeeded with proven checkpoint number ${monitor.provenCheckpointNumber}`);
+  });
 });
