+++ yarn-project/end-to-end/src/multi-node/single-node/proof_fails.parallel.test.ts
@@ -17,17 +17,17 @@ import { RootRollupPublicInputs } from '@aztec/stdlib/rollup';
 
 import { jest } from '@jest/globals';
 
-import type { EndToEndContext } from '../fixtures/utils.js';
-import { MultiNodeTestContext } from './multi_node_test_context.js';
+import type { EndToEndContext } from '../../fixtures/utils.js';
+import { SingleNodeTestContext } from '../single_node_test_context.js';
 
 jest.setTimeout(1000 * 60 * 10);
 
-// Suite: 2 parallel scenarios testing proof-submission failure paths. MultiNodeTestContext with single
+// Suite: 2 parallel scenarios testing proof-submission failure paths. SingleNodeTestContext with single
 // sequencer node, no initial prover (prover nodes created in test bodies). Timing: ethSlot=8s,
 // aztecSlot=2×8=16s, epoch=8, proofSubmissionEpochs=1 (default), blockDurationMs=3s,
 // cancelTxOnTimeout=false, inboxLag=2 (v5 always enforces the timetable, so the former enforceTimeTable
 // override is gone). Prover Delayer steers proof tx timing.
-describe('multi-node/proof_fails', () => {
+describe('multi-node/single-node/proof_fails', () => {
   let context: EndToEndContext;
   let l1Client: ViemClient;
   let rollup: RollupContract;
@@ -38,10 +38,10 @@ describe('multi-node/proof_fails', () => {
 
   let L2_SLOT_DURATION_IN_S: number;
 
-  let test: MultiNodeTestContext;
+  let test: SingleNodeTestContext;
 
   beforeEach(async () => {
-    test = await MultiNodeTestContext.setup({
+    test = await SingleNodeTestContext.setup({
       maxSpeedUpAttempts: 0, // No speed ups
       startProverNode: false, // Avoid early proving
       ethereumSlotDuration: 8,
@@ -49,7 +49,6 @@ describe('multi-node/proof_fails', () => {
       aztecSlotDurationInL1Slots: 2,
       blockDurationMs: 3000, // 3s blocks → 2 blocks per checkpoint under pipelining
       cancelTxOnTimeout: false,
-      inboxLag: 2,
     });
     ({ context, l1Client, rollup, constants, logger, monitor } = test);
     ({ L2_SLOT_DURATION_IN_S } = test);
